+++ allauth/idp/oidc/internal/oauthlib/request_validator.py
@@ -338,9 +338,7 @@ class OAuthLibRequestValidator(RequestValidator):
             )
         )
         adapter.populate_id_token(id_token, request.client, request.scopes)
-        jwk_dict, private_key = jwkkit.load_jwk_from_pem(
-            adapter.get_current_private_key()
-        )
+        jwk_dict, private_key = jwkkit.load_jwk_from_pem(adapter.get_signing_key().pem)
         return jwt.encode(
             id_token, private_key, algorithm="RS256", headers={"kid": jwk_dict["kid"]}
         )