+++ ChangeLog.rst
@@ -12,7 +12,6 @@ Noteworthy changes
 -------------------
 
 - IdP: Added support for Client ID Metadata Document (CIMD).
-- IdP: Added key rotation support and cache control for the JWKS endpoint.
 
 - IdP: Added ``IDP_OIDC_REFRESH_TOKEN_EXPIRES_IN``, allowing refresh tokens to
   expire. Combined with ``IDP_OIDC_ROTATE_REFRESH_TOKEN``, this results in a
@@ -23,6 +22,9 @@ Noteworthy changes
 
 - IdP: Added an ``oidc_cleartokens`` management command that deletes expired
   OpenID Connect tokens.
+
+- IdP: Added key rotation support via ``IDP_OIDC_PRIVATE_KEYS``, as well as
+  cache control for the JWKS endpoint.
 
 
 Fixes